clippy fixes

+++ pallets/schema/src/weights.rs
@@ -15,8 +15,8 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	fn create() -> Weight {
 		Weight::from_parts(522_000_000, 0)
 			.saturating_add(Weight::from_parts(20_000, 0))
-			.saturating_add(T::DbWeight::get().reads(2 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(T::DbWeight::get().reads(2_u64))
+			.saturating_add(T::DbWeight::get().writes(2_u64))
 	}
 }
 
@@ -25,7 +25,7 @@ impl WeightInfo for () {
 	fn create() -> Weight {
 		Weight::from_parts(522_000_000, 0)
 			.saturating_add(Weight::from_parts(20_000, 0))
-			.saturating_add(RocksDbWeight::get().reads(2 as u64))
-			.saturating_add(RocksDbWeight::get().writes(2 as u64))
+			.saturating_add(RocksDbWeight::get().reads(2_u64))
+			.saturating_add(RocksDbWeight::get().writes(2_u64))
 	}
 }
